requirements.txt
neuroconv==0.4.6
spikeinterface>=0.100.2
nwbwidgets==0.11.3
nwbinspector==0.4.33
ndx-events==0.2.0
ndx-photometry>=0.4.0
tdt==0.6.5
PyYAML==6.0.1
openpyxl==3.1.2
pandas>=2.2.0